requirements.txt
pymupdf
pandas
scikit-learn
nltk
torch
torchvision
transformers
faiss-cpu
matplotlib